# docs/swagger/collections.yaml
openapi: 3.0.0
servers:
  - description: Listmonk Developement Server
    url: http://localhost:9000/api
info:
  version: "1.0.0"
  title: Listmonk
  description: The API collection for listmonk
  license:
    name: AGPL-3.0 license
    url: https://github.com/knadh/listmonk/blob/master/LICENSE
...
paths:
...
  /lists:
    get:
      description: retrieves lists with additional metadata like subscriber counts. This may be slow.
      operationId: getLists
      tags:
        - Lists
      parameters:
        - in: query
          name: page
          description: total number of pages
          required: false
          schema:
            type: integer
        - in: query
          name: per_page
          description: number of items per page
          required: false
          schema:
            oneOf:
              - type: integer
                description: Number of items to return per page
              - type: string
                enum: ["all"]
                description: Return all results without pagination
        - in: query
          name: query
          description: Optional string to search a list by name.
          required: false
          schema:
            type: string
        - in: query
          name: order_by
          description: Field to sort results by. name|status|created_at|updated_at
          required: false
          schema:
            type: string
            enum: ["name", "status", "created_at", "updated_at"]
        - in: query
          name: order
          description: ASC|DESC Sort by ascending or descending order.
          required: false
          schema:
            type: string
            enum: ["ASC", "DESC"]
        - in: query
          name: minimal
          description: When set to true, returns response without body content
          required: false
          schema:
            type: boolean
        - in: query
          name: tag
          description: Tags to filter lists. Repeat in the query for multiple values.
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: list of metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/List"
                      total:
                        type: integer
                      per_page:
                        type: integer
                      page:
                        type: integer

    post:
      description: handles list creation
      operationId: createList
      tags:
        - Lists
      requestBody:
        description: new list info
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewList"
      responses:
        "200":
          description: updated list object
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/List"
...
components:
  schemas:
...
    List:
      type: object
      properties:
        id:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
        uuid:
          type: string
        name:
          type: string
        type:
          type: string
        optin:
          type: string
        tags:
          type: array
          items:
            type: string
        subscriber_count:
          type: integer
        default_messenger:
          type: string
          description: "Messenger name to use as default for this list"
        description:
          type: string
    NewList:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [public, private]
        optin:
          type: string
          enum: [single, double]
        tags:
          type: array
          items:
            type: string
        default_messenger:
          type: string
          description: "Messenger name (email only) to use for automatic emails involving this list"
        description:
          type: string
...